remove outdated/unneeded Crypt::OpenPGP requirements checking
* use only Crypt::OpenPGP (or `gpg`) as requirement for installation
* fixes #14
* fixes rt-bug#108377 (https://rt.cpan.org/Ticket/Display.html?id=108377)

.# Discussion

The previously listed requirements for Crypt::OpenPGP were outdated and, unnecessarily,
blocked installation on more recent MSWin32 64-bit perl versions. Math::Pari, and
other modules relying on Math::Pari, being the problem, were removed as a dependecy as of
Crypt-OpenPGP-1.12.

[ref] https://rt.cpan.org/Ticket/Display.html?id=108377 @@ https://archive.is/BnRga

As a replacement, simply specifying Crypt::OpenPGP (or a working `gpg` executable) as the
prerequisite for this modules should suffice.