prefer gpg2 when possible
On systems that have both 'gpg' and 'gpg2', 'gpg' will sometimes be gpg
1. Prefer 'gpg2', which is more likely to be the newer version.

This can be important, because gpg 1 does not include the key
fingerprint when generating a signature, instead only including the key
id. This is true for all versions of gpg prior to 2.1.16. And gpg
versions 2.3.0 and newer only support the --auto-key-retrieve option
when the signature includes the fingerprint.