Fixes #36 auto-key-retrieve is obsolete
This patch reproduces the same functionality as auto-key-retrieve
If verify fails it gets the key ID from the error message
and retrieves the key from the configure key server.
It then attempts to verify the signature again.

Auto retrieval of the key that created a signature and then
essentially auto trusting it for verification.

You should only download keys you trust and import and trust them
prior to verifying a module.  Admittedly a poor recommendation for
this module.